Removed the experimental code with LoggingWrapper
The idea was to be able to call LoggingWrapper.debug('msg') from
any module so that the wrapper would inherit the Logging class.
This turned out to be too complicated for the current skill set.
Not sure if this is feasible in a such way that allows code
without any 'tricks'.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>