Added conditional code profiler with cProfile
There is now experimental profiler integration with the cProfile.
The intention was to use argparse to read the hidden '--profiler'
switch but the problem was that this information was available
too late.

The profiler needs to be enabled as soon as possile in order to
catch the tool startup. Because of this, the --profiler switch is
read directly from the system arguments and at the momemnt the
--profile switch from Config() is not really used. This is just a
placeholder and "documentation" for the switch.

There was also experimental with the pyflame, but that profiler
would have require constant running service which PID stays the
same. There are instructions also how to install the pyfleme.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>