Cleaned pylint warnings from missing documentation from tests
Also disabled the no-self-use check in pylint. The reason was
that the tests are made from classes and each test is executed
as a method by pytest so there is no one calling the tests from
the implementation.

Another question is that could/should the tests be functions
instead of methods in a class that is never called. Currently
the thought is something like "class/OO stuff is used in the
main code so also in tests". Not sure if the logic can hold.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>