Updated documentation to use terms 'resolution' and 'resolve'
The troubleshooting reference solutions were previously called
logs. This was considered misleading and better terms resolve
and resolutions are used for this.

The resolve and resolution is intended to mean a longer text
file or a string that contains multiple command examples and
more detailed description about the problem that was solved.

On top of resolutions, the intention is to store snips or
snippets that are mostly one line command examples.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>