Fixed pylint warning from code similarities in tests
The test code contain setup and teardown that are same for
two tests. Pylint complains about these two methods in two
modules. The setup and teardown are needed because they are
called by the test framework and cannot be removed.

It seems that pylint 1.7.2 has open fault that makes disabling
of the check in portion of the code not to work /1/.

It also seems that pylint 1.7.2 the ignore-comments and
ignore-docstrings do not seem to work as expected. Even
when the values should state that these are ignored, the
min-similarity had to be increased to 8 because there are
7 lines in these two referred methods.

The target is to get the warnings down to see changes
easily. This fix could be fixed later on by setting the
min-similarity to 4 and use the ignore docstring and
comment settings combined with the code tag to ignore
the duplicated code.

/1/ https://github.com/PyCQA/pylint/issues/214

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>