Added possibility to use sqlite3 as in memory database
There is also a dummy test for testing the Storage class. The
though is that testing might be easier if the file is not
created for the database. For example parallel running of the
database tests might be possible. Currently the database is
set as a shared memory database. It is to be seen if this needs
to be changed or not.

A lot of work is still to be done to have the Snippet to be
added into the storage.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>