Added basic insert and regexp based select from database
The regexp based select works based on keywords and there is a
single test for it. A quick integration of the Storage() to
store and insert the parameters from command line was added.

Adding the same snippet for the second time kicks the unique
constraint so all the error and special case handling is
missing.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>