Added yaml file export for snippets
The snippets can be now exported to yaml file as defined
byt the user. The command example is like below which
creates a snippy.yaml file in the same folder where the
command is run.

The file extension must be either yaml or yml. The tool
uses always yaml so that the user file extension '*.yml'
is always converted to '*.yaml'.

    $ python snip.py -e snippy.yaml

There are likely shortcomings in the special character
conversions. The yaml file generation may need refactoring.

The reason to use the string based manual formatter is to
try to keep the format exactly to allow style that pleases
authors eye. This might be problematic in case of special
characters which is to be seen.

In case the tool spots a yaml file error when the generated
string is tested with yaml.load(), the tool exits with fatal
failure as of now.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>