Added support for multiline snippets
The snippets can be now added through editor with -w|--write
option. This was considered the easiest way to provide this
to end user.

The command line switch for adding snippes does not support
splitting the snippet into multiple lines at this point of
time.

The multiline snippets are output as a string with newlines
in exported yaml or json files. Same behaviour is with the
console output.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>