Fixed tags and find arguments tests and prints
The tags and find arguments are provided as list from argparse
because of nargs='*'. The default value for these arguments
was set to empty string when it should have been an empty list.
This was incorrectly tested and printed.

Fixed the argument and configuration printing formats for
other variables that are not string.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>