Added support for text file export
The -e|--export argument now recognizes *.txt or *.text file
extension and creates a *.txt file into file and location
specified by user. The format is currently the same as with
the terminal output.

Future improvement might contain categories where the snippets
could be organized for example in categories: linux, docker
or OpenStack.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>