Changed snippet() and resolve() classes to initialize with storage
The storage object is currently created at the main level. There
are no other reasoning that that this "feels a high level init
task" and thus it is passed to snippet and resolve management
objects rather than created there.

The change moved the storage parameter from run to init method
since the later was felt more logical place.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>