Changed log level management
There is now a hidden --debug switch that enables logger.DEBUG
log level. By default the log level is logger.ERROR.

The log handler hierarchy was also taken into use. Previously
all the modules used a root level logger. Now each module
creates a logger with 'snippy.' combined with the module name.
This allows more sophisticated log level control on module
level.

The log level propagation relies on that the module level logger
does not set the log level. It was supposed to be possible to
set the module level log level to logger.NOTSET and but it was
not working as expected. The level must be completely left out
from logging.basicConfig.

Also the application log level is set to root level. It was
tried to be set with logging.getLogger('snippy') but that was
not working as expected as of now.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>