Added support for categories
The snippets can now be categorised. User can give a single
category for each snippet. The intention is to allow grouping
of the data into larger groups. This should allow for example
directory or text block based separation of the data.

The intention is to build command reference sets for example
for Linux, Docker or OpenStack and be able to sort and print
these sets sequentially.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>