Changed most of the command line arguments
The command line arguments were changed in try to make them
more readable and logical when supporting both snippets and
resolves.

There is now mandatory job role 'snippet' or 'resolve' which
is needed to define where the data is stored. Each role has
same set of of tasks like create, search or delete.

User can use supplementary services like give the input with
an default editor or use file input.

The ID allows identifying the target row/record in the database.
Currently this is only for delete.

The optional parameters can be defined now for snippets. The
thought is that user would be more comfortable with editor
or input files than with a full command line argument list
to create a new snippet.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>