Added generation of has for each snippet
The snippets can be now identified based on message digest.
Intention is to operate the snippets based on this identity
instead of the database row number. This also allows detecting
corrupted data for example during import and merge operations.

The delete job still works based on database row number and
there is now check on uniqueness of the hash.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>