Changed storage initialization to allow one liner
The Storage() now returns self in order to allow calling
the init() method in the same line when the Storage object
is created.

Currently it feels that the additional init() on top of
__init__ is unnecessary. The init(), which creates the
database is always called and there is no current need
to have the separation.

The init() was not yet merged to __init__ even though
there is no clear reason to keep it like this.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>