Changed command line arguments
The refactoring to improve the command line arguments is
still ongoing. The new changes are still in comments. Now
the command layout in help is looking like what it was
designed originally.

There were some changes made to original design because
of Python argparse limitations. It could have been possible
to overcome these by accessing the internal methods from
argparse but it was not considered that this is not good.

The version information was also changed to use the
metadata from python setuptools.

For some reason the setuptools metadata contains only the
version even though there are other attributes as well in
the setup.py file.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>