Changed the command line arguments
There were minor changes on top of previous refactoring of
the command line interface. The changes are still in comments
but the actual argparse and user interface should be ready
to take into use.

There were few functionalities added to the CLI interfaces
to support proper feature set for first release. For example
the import templates and the --template option does not
have implementation yet.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>