Changed command line interface
The new command line interface look and feel was taken
into use. The target was to make it more like a Linux
command line examples. Author's view was that the default
argparse help was too cluttered and hard to read.

Now the commands should contain enough parameters to
support first version. Some of the commands like the
templates are not yet working.

With the same change there were minor refactoring
made in order to try to improve the code readability.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>