Fixed fatal failure when creating duplicated content
The case where used added duplicated content was handled
previously with exception and exit. This is now gracefully
handled.

There are two use cases to handle this. User gets a failure
prompt that indicates the duplicated example and its message
digest. User may change the create operation to update and
define the message digest for the update.

User may also just replace the creation operation to update
operation and launch a vi editor that contains the data read
from command line as a default. Then user may edit the
content and save the file to get changes updated.

There is now new documentation chapter that describes these
two use cases. The intention is to document the supported
and tested use cases because these are eventually getting
more complicated and hard to remember.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>