Added exit code output at the end of execution
By default the logging is off and user is supposed to get only
OK or NOK indication of the operation. The logging can be
enabled with -vv or --debug options. Currently the exit code
is set only from one place.

The intention is that the option -q would suppress the exit
code but this is not yet done.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>