Fixed incorrect message digest for duplicated content
The duplicated content message digest was printed from the
new data. This caused incorrect message digest because only
the content is checked with unique constraint. If the other
data is different, it will create different message digest.

Now the printed message digest matches to the one that is
stored and needs to be updated.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>