Fixed updating content with editor
When content was updated with editor, there were two faults.
The tag list from the editor was processed to a single string
in a list even though there were multiple tags.

The second problem was that in case the field was empty, an
empty string was stored in a list. For example no links defined
in the editor caused the list to contain one item which was
empty string.

Now when there are no changes made in the update, the digest
remains the same as before as it should.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>