Added support to print all snippets on console
The '--sall .' format with the dot allows matching all the
records.

The search query for the --sall, --stag and --sgrp is based
on regexp but the only special character passed as of now is
the dot. The reason is that used could provide arbitrary
string that could not be valid regexp which would cause an
exception in current implementation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>