Changed snippet console print colors
The colors in the console print were slightly changed. Intention
was to highlight the actual command more than the supplementary
information like links and tags.

Added also example how to grep the commands from search result.
The color change works better in this case as well because the
output is more lighter making the command more visible.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>