Changed exit cause print to include also log printing
In order to reduce logging lines in code, the cause setting
and log printing was combined.

There was pattern emerging that when an exit cause was set,
similar log was also printed. Now the informative log is
printed at the same and the logging can be done from the
main level with single line.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>