Changed how the version string is read
There is now global file called version.py that defines the
version string for the setuptools and --help option.

The previous alternative relied on pkg_resources package which
was failing when the tool was not installed by pip because this
package searched the installed Python packages. This is likely
the common pattern to do this installation but as of now, it
was thought that the tool should not throw an exception if the
'make install' or 'pip install .' was not done.

Python seems to have multiple ways of doing this and all the
solutions seem somewhat awkward at the moment. More information
for example from /1/ and /2/

/1/ https://packaging.python.org/guides/single-sourcing-package-version/
/2/ https://stackoverflow.com/questions/458550/standard-way-to-embed-version-into-python-package

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>