Fixed build and failing tests
The snippet management was refactored. The snippets are now
managed as tuples instead of hashes. The logic still may need
some tuning but the intention is that the database and storage
layers would take care of the conversion. The purpose is that
the users from storage layer would always get an constant view
of the data. And the data is now moved everywhere in immutable
tuples.

For example /1/ (old post) mentions that the Python DB API
fetchmany() results a list of tuples.

/1/ http://news.e-scribe.com/397

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>