Changed default shown in group field in editor template
Previously the 'default' value for the group was shown to
the user when new content was added with editor. This was
considered confusing and no value is shown in the group
field if the value is the default one.

If user defines the --group option, that value is normally
shown in the editor in the default template.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>