Added first workflow test case to verify snippet creation
There were minor refactoring needed to get this smoothly from
test case point of view. The Snippy() class now supports
returning the storage that allows verifying the data from
the tool storage.

The same exposing could allow for example verifying that
the configuration parameters are correct.

The automatic documentation for the workflow tests is not
clear yet. There is initial draft about the documentation
string for the workflow tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>