Fixed failing database tests when run individually
The previous change broke database test when they were run
individually. The problem was that the removed database.sql
path mock caused a situation where the Config() was not
initialized and thus the real call to method was not working
because the data structures were not initialized in Config().

But the question is that why the Config() gets initialized
in the first place when all the tests are run with make test?
It may have to do something with the folder and/or command
differences.

Tried also do slow initializers for Storage() and Config()
but they turned more complicated then though. The intention
was to be able to control the database-in-memory setting
before the actual database would be created. This would
require removing the initializer of Storage() away from
Snippy() initializer and have similar two phase Config()
in order to support nice setUp and tearDown in tests.

Intention was like below but the problem is that the sys
args get consumed when Snippy() is created. The input args
need to be set in each test separately.

    def test_creating_new_snippet_from_command_line():

        sys.argv = ['snippy', 'create'] + Snippet().get_command_args(0)
        self.snippy.run()
        references = Snippet().get_references(0)
        Snippet().compare(self.snippy.storage.search(digest=references[0][DIGEST])[0], references[0])
        Snippet().compare(self.snippy.storage.search(content=references[0][CONTENT])[0], references[0])
        assert len(self.snippy.storage.search(digest=references[0][DIGEST])) == 1
        assert len(self.snippy.storage.search(content=references[0][CONTENT])) == 1
        self.snippy.release()

    # pylint: disable=duplicate-code
    def setUp(self):
        """Setup each test."""

        self.snippy = Snippy()
        self.snippy.config.set_storage_in_memory(True)

    def tearDown(self):
        """Teardown each test."""

        Database.delete_all_snippets()
        self.snippy.disconnect()

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>