Fixed generated document license layout and version string
The generated rst document used the license file and formatted the
text. Now the license file is pure text.

The version string for the documentation is read now from one place
from the version.py file.

The HTML CSS folder _static for Sphinx is now configured to static in
order to avoid problems with git and folder names with underscores.

Signeaoff-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>