Changed storage location
The Sqlite database is now stored under the project. This
makes it easier and more logical to install with pip. There
is also additional data from the default data which intends
to store default imports for the snippets and resolutions.

Fixed the pip packaging by removing accidentally included
tests.testlib

Changed the order of links and tags in the console print.
It was considered to be more readable like this especially
when there are more than one link or the links are very long.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>