Changed the snippy.db file to be ignored by git
The snippy.db is intended just to be empty for the Python packaging.
In order to get the database file removed when uninstalling, the
file must exist. The development changes into snippy.db are not
intended to be commit into repository.

The snippy.db file was marked with command below not to be tracked:

  $ git update-index --assume-unchanged snippy.db

The command can be reverted with:

  $ git update-index --no-assume-unchanged snippy.db

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>