Changed content, tags and link management to tuples
The content is now managed as a tuple where each row is an
element in a tuple. The tags and links were previously lists
but they are now converted to tuples since there were no
reason to mutate these fields.

The reason for the content to be in a tuple was to get better
format in the exported yaml and a thought that processing the
snippets would be cleaner when the content is in array.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>