Added delete to work with content
It is now possible to delete a snippet by defining the
-c|--content option. The content has to match exactly
in order to run the delete operation.

The delete was also fixed to be more careful about the
provided digest. An empty or too short (less than 16
characters) digest considered valid anymore. The reason
is that in case there is one snippet in the database,
it gets delete if user accidentally set an empty digest.

The search operation was fixed in case there is not search
keywords at all. Previously there was exception because of
variable being accessed before it was defined.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>