Changed TODI list and removed the proposal to rename CONTENT
The initial thought was that the content[Const.CONTENT] would
be not so good because the content word was repeated. The idea
was to rename the CONTENT to DATA.

After trying the implementation, it was noticed that then the
command line interface -c|--content would be a bit misleading
and the -d|--digest was taken. Also the database schema would
need to be synchronized.

Then the conclusion was that the content referring to content
was OK since it is the content :)

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>