Added --all category for search operation
It is now possible to use --all category but only with search.
The reason is that the --all does not make sense with other
than the search.

There might be a use case with export --all but as of now it was
left out because it was considered that dumping both content
categories into same file would not be nice for example in case
of importing.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>