Changed content category internal management
The content category terminology is now harmonized to use the same
term 'category' everywhere in the code. The category is also part of
the content definition in the SQL tables. In case of create, update
and import operations, the content provides the category inside the
data structure. In case of rest of operations, the content category
is given into Storage(). The intention is to improve readability and
to tell what is changed instead of hiding the content reading into
lowest Database() layer.

This allows for example combining the snippets and solutions into
one list and print the proper content according to the category.

The run method on Snippy() was renamed to run_cli. In future there
might be for example REST API that does not use the CLI.

The file name was added to the database definition in order to bring
the solution file name into database as well.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>