Fixed filename update in case of solution content
The filename in solution was previously read from the content
data since the field was not in the database. The field was
moved previously into database where the updated value is
stored. Changed the code to use the database value instead
of searching it from the content data.

Two other minor code layout change included as well.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>