Changed digest management in update operation
The update operation carries the old digest in the content
and it is not passed as parameter to update method anymore.
Only the new digest is passed as separate parameter to
database update.

This harmonizes the code with insert and clears the code a
bit by removing extra parameter from the method. The edited
content and the Config().get_content circulates the values
that cannot be edited from original content to updated
content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>