Changed snippet update operation internal editor flag
The editor flag was previously forced for update operation
with digest only. The update operation is possible also
based on content which did not have enforced editor setting.

The editor flag is not forced anymore in Config(). The flag
can be controlled from Snippet() or Solution() when content
is created.

This was misalignment due to code changes that is now fixed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>