Added support to search message digest with --sall option
The --sall option now searches also the message digest which
allows output of a single content based on message digest or
part of it.

Note that the terminal output shows only 16 characters from the
message digest but the search is made from full 64 character
digest. This may produce results that match to a partial digest
that is now visible on the terminal output

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>