Changed logs that formatted the string always
Following best practises not to format the log string with
variables all the time. The syntax '.format()' forces the
log string intepolation even when the log is not printed
due to log level.

This change contains the update for few remaining logs after
the refactoring of this topic in the past.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>