Fixed unhandled exception in case of json load or dump failure
The import for yaml and json is made in the last moment before
just using the load or dump in order to speed up startup time.
Both json and yaml management is in the same try/catch block.
Previously the yaml.YAMLError was the first exception type
to be checked. If the failure was in json side, the yaml module
was not loaded and this caused unhandled exception. Now the
order of exception types for Migrate() load and dump is
significant.

The Python generic TypeError and ValueErrors raised by must be
first in order to handle the json exceptions.

The import statements could have been moved outside of the
try/catch block but the thought was to import these modules
only when needed in order to try to maintain responsiveness
of the tool

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>