Changed how the editor is selected for create and update
The previous way of blindly reading the EDITOR environment
variable is likely not safe from security point of view. It
is not fully sure how to safely read and check environment
variables in OS independent way. There is likely differences
between Linux, Windows and Mac that are not fully understood.

Because of this, the editor is enforced as vi as of now.

The Cause() management was also refactored due to this change.
The problem was that the cause code setting was in Config()
which did not allow the usage of the set_cause from every
module due to circular import dependencies. Because of this,
the Cause() class was created to be used everywhere.

The cause code management change also tries to create code
where the cause is set and empty values are returned in
case of failures which should be gracefully handled. This
tries to avoid endless checking of the failure codes. Code
just flows to the end with empty values and prints the
origin failure cause.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>