Changed cause management
Cause text can be now set from any module and it will persist
the first cause set which is printed at the end of the tool
execution. The intention is to provide clear text format print
for end user to allow if the operation failed and for which
specific reason.

There is no other causes than text based. The cause code was
removed since it was too "big" for small tool like this.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>