Changed test case names
Test cases are now divided into UT (unit test) and WF (workflow)
types of tests. The workflow tests test the whole application and
the unit tests test a single class.

The naming was also simplified for now just to refer to test
type, tested class and operation in case of unit test and
operation and category in case of workflow tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>