Added support for -d|--digest option with search operation
The search operation now supports also digest on top of
keywords and content data like:

    $ python snip.py search -d 53908d68425c61dc

Added option --no-colors to disable colors from terminal
output. This was intended for testing to remove the color
characters to make the search string comparison easier.
This is also left out from command line help to keep the
help text short and easier to read.

The --no-color was added in previous commit but it was
left out from commit log by mistake.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>