Changed printing content when filter option was used
Removed the logic to strip out ASCII color codes from the
terminal print because if was not needed. The text is now
formatted always without colors if the --filter option is
used. This allows the filter regexp to work as expected
because the color codes within the text should not affect
to the filter itself.

The functionality changed only internal and there are now
external changes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>