Fixed failure message in import and category change in update
1. Exit cause mismatch in case if import operation

In case the content data already existed in the database when it
was imported, it previously generated NOK exit cause. This is now
considered as OK case since the content exist there.

This means that if other fields are changed, the content in the
database does not get changed.

The existence of this condition is visible only from logs.

These are intended behaviours.

2. Category change in case of update operation

If user defined the content with message digest and used category
option that did not reflect the updated content, the category was
changed and the update content was stored with incorrect category.

This is now corrected so that the category does not change for the
updated content even when user provides incorrect category.

3. Default content update

Updated also the default solution and snippet content because
they had incorrect message digest.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>