Fixed failing delete workflow test in Python2.7
The delete workflow test case was passing in Python3 but not in
Python2. The reason is that in case of Python3, the database is
in memory but in case of Python2 it is file. The Python2 does
not support in memory Sqlite3 database.

The Python2 database file is currently same for every test case,
which also means that parallelism does not work currently in
Python2. If there is one failed it seems that the Python2 related
database helper code fails to clear the situation

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>