Changed --no-colors option to --no-ansi
The option prevents using ANSI control characters like color
codes in the terminal output. This is intended to be used
with post grep filters where the ANSI characters make the
regexp no to behave like they should because of hidden
characters.

Fixed also example help that contained incorrect option name
for the --no-ansi.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>