Fixed export and import operations with non-existent file
Export failed to handle a case where the file was not specified
with -f|--file option. If user now leaves the -f option away, it
defaults to either snippets.yaml or solutions.yaml based on the
operation category.

    $ snippy export # Exports all snippets to snippets.yaml.
    $ snippy export --solution # Export solutions to solutions.yaml.

Fixed also import operation that was using non-existent file which
caused exception. Now the non-existent file case causes proper cause
for the end user.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>