Changed default content installation to use quiet mode
The docker compilation and snippy installation does not print
'OK' after importing the default snippet and solution content.
The plain 'OK' was considered misleading without any additional
text and it was considered better to reduse the verbosity of
compilation.

The test suite should take care that the default imports
are working.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>