Added importing of text file based template
It is now possible to import from text file based templates.
The import supports only importing of one content and it is
not intended for data migration like JSON or YAML based
operations.

Currently the content category (snippet or solution) must
be specified according to the imported content.

You can now use:

    $ snippy import --solution --file solution.txt

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>