Changed argument processing for template file
The --template parameter always expects only one argument. It is
not possible to get pass of argparse without any argument or with
multiple arguments to --template option.

Because of this reason, internal logic build similar to keyword
management is unnecessary with the --template option. The parameter
is None when it is not set and this can be used to determine when
the option was set.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>