Changed migrate operation arguments for defaults and templates
The content defaults can now be migrated with option --defaults
instead of magical filename 'defaults' to -f|--file option.

  $ snippy import --snippet --defaults
  $ snippy import --solution --defaults
  $ snippy export --solution --defaults

The templates are now handled in same manner than defaults.
The template can be exported with option --template. The import
can be done with same --template option or by specifying the
filename like in examples below.

  $ snippy export --solution --template
  $ snippy import --solution --template
  $ snippy import --solution -f solution-template.txt

The intention was to avoid magical filenames that have special
meaning. Also this harmonizes the usage of migrate options
between templates and defaults.

The migration option group is now mutually exclusive so it is
not possible to combine -f|--file, --template or --defaults
in the same command.

Removed also the -f|--file from the editing options.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>