Fixed importing content with conflicting category
Previously when user imported a text based template, the category
defined in command line was overruling the content in the template.
This created a fault where solution content was stored in snippet
category. Now the when the data is imported from text template,
the category is automatically verified and updated if needed.

This means that user can just import the file without considering
what is the content category.

The text based import is intended to import a single content.
JSON and YAML based content is intended to manage the data
migration and processing longer content lists.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>