Fixed crash when content template was imported without modifications
In case due to some internal error that lead to content data being
empty, the sqlite3 search query could have been left empty. The
reason was that none of the internal search criteria matched.
This lead to exception in the select statement.

Improved also user failure cause to indicate if no content was
imported at all.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>