Added support to specify migrated content with message digest
It is now possible to export a content with digest like:

    $ snippy export -d e95e9092c92e3440

The command is category independent and the resulting file
name is based on internal defaults or the filename setting
that is available in solution category.

It is now also possible to import (update) content with
specific message digest. This allows updating existing
content like:

    $ python import -d e95e9092c92e3440 -f howto-debug-elastic-beats.txt

Importing a new content from file works like before:

    $ python import -f howto-debug-elastic-beats.txt

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>