Fixed exporting solution to defined file
In case a solution was exported to a text file, it always used
the filename defined in solution even though when the user
defined the output filename from command line. The user defined
file from command line is always honored.

If user does not define the filename, the name is taken from
the content or from a default value.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>