Fixed parsing of filename from solution text template
In case the filename was not defined in the template, parsing
of the filename failed and resulted unexpected name.

Now the filename can be properly left empty in the template
in which case the tool will use default name for the exported
solution if user does not provide target filename from command
line.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>