Changed pylint checks on similarities in test code
The pylint complains similarities in the test code that
sets the solution data. This data has similarities, but
it is not the same.

Disabling this warning with # pylint:disable=duplicate-code
was not working. The reason is unknown. It was suspected
that the mentioned file in the warning could be wrong.

Because of these reasons, the test case similarities
minimum line count was set to 15 which covers the solution
header and the first few similar lines.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>