Fixed test case reference with Python2 and packaging
The test case reference command did not work with Python2
because of file line reading differences between Python3
and Python2.

Improved the failure handling of test case reference
printing with packaging. The test cases are not currently
package with the release and documentation cannot be
printed in distributed version. Now the error handling
for not finding the tests at all is smooth.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>